Move CSV headers to a separate file.
